+++ src/components/NewsList/NewsList.css
@@ -62,34 +62,3 @@
 
   padding: 24px;
 }
-
-.newsButton {
-  width: 24px;
-  height: 40px;
-
-  border: none;
-  padding: 0;
-
-  background: none;
-}
-
-.prev {
-  position: absolute;
-  top: 50%;
-  left: -47px;
-
-  transform: translate(-50%, 50%);
-}
-
-.next {
-  position: absolute;
-  top: 50%;
-  right: -47px;
-
-  transform: translate(50%, 50%);
-}
-
-.newsButton > img {
-  width: 24px;
-  height: 40px;
-}
